requirements.txt
allure-pytest==2.9.45
pytest==6.2.3
selenium==4.1.0
requests==2.26.0
pytest-xdist==2.2.1
